requirements-dev.txt
Flask-Script==2.0.5
pep8
